refactor: move API tools to internal tool registry, extract API tool logic to helpers
