+++ x-pack/platform/plugins/shared/agent_builder/server/plugin.ts
@@ -229,7 +229,7 @@ export class AgentBuilderPlugin
       searchInferenceEndpoints,
       security: securityPlugin,
     } = startDeps;
-    const { elasticsearch, security, uiSettings, savedObjects, dataStreams, featureFlags } =
+    const { elasticsearch, http, security, uiSettings, savedObjects, dataStreams, featureFlags } =
       coreStart;
 
     this.cleanupLegacySmlTasks(taskManager).catch((error) => {
@@ -241,6 +241,7 @@ export class AgentBuilderPlugin
       security,
       securityPlugin,
       elasticsearch,
+      http,
       inference,
       spaces,
       actions,
