+++ x-pack/platform/plugins/shared/agent_builder/server/services/execution/run_agent/api/index.ts
@@ -1,0 +1,23 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { targetSchema } from './types';
+export type {
+  ApiRegistry,
+  ApiRegistryDefinition,
+  ApiRegistryMeta,
+  ApiRequest,
+  ApiTarget,
+  LoadedApi,
+} from './types';
+export { loadApi } from './load_api';
+export { listApisForTarget } from './list_apis';
+export type { ApiSummary } from './list_apis';
+export { toDescribedSchema } from './describe_schema';
+export { getUnsupportedReason, prepareApiRequest } from './prepare_request';
+export type { PrepareApiRequestFailure } from './prepare_request';
+export { dispatchApiRequest, getFailureDetails } from './dispatch_request';
