+++ x-pack/platform/plugins/shared/agent_builder/server/services/execution/run_agent/tools/api/index.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { createDiscoverApisTool } from './discover';
+export { createDescribeApiTool } from './describe';
+export { createExecuteApiTool } from './execute';
