+++ x-pack/platform/plugins/shared/agent_builder/server/services/execution/run_agent/tools/register_internal_tools.ts
@@ -24,6 +24,7 @@ import { createAskUserQuestionTool } from './ask_user_question';
 import { createReadFileTool } from './read_file';
 import { createListFilesTool } from './list_files';
 import { createBashTool } from './bash';
+import { createDiscoverApisTool, createDescribeApiTool, createExecuteApiTool } from './api';
 import { createTodoTool } from '../../../tools/builtin/todo';
 import { builtinToolToExecutable } from '../utils/select_tools';
 import type { BackgroundExecutionService } from '../background_execution_service';
@@ -74,6 +75,7 @@ export const registerInternalTools = async ({
     filesystemService,
     bashService,
     todoStateManager,
+    selfClient,
   } = context;
 
   const interactive = executionMode !== AgentExecutionMode.standalone;
@@ -90,6 +92,13 @@ export const registerInternalTools = async ({
   // Todos — FF-gated.
   if (experimentalFeatures.todos) {
     tools.push(createTodoTool({ todoStateManager }));
+  }
+
+  // HTTP API introspection/invocation — FF-gated.
+  if (experimentalFeatures.apiTools) {
+    tools.push(createDiscoverApisTool());
+    tools.push(createDescribeApiTool());
+    tools.push(createExecuteApiTool({ selfClient }));
   }
 
   // Sub-agent + sleep — experimental, and not available in standalone mode.
