+++ x-pack/platform/plugins/shared/agent_builder/server/services/execution/runner/run_agent.ts
@@ -37,6 +37,7 @@ export const createAgentHandlerContext = async <TParams = Record<string, unknown
     request,
     spaces,
     elasticsearch,
+    http,
     savedObjects,
     modelProvider,
     toolsService,
@@ -74,6 +75,7 @@ export const createAgentHandlerContext = async <TParams = Record<string, unknown
     logger,
     modelProvider,
     esClient: elasticsearch.client.asScoped(request, { projectRouting: 'space' }),
+    selfClient: http.selfClient,
     savedObjectsClient: savedObjects.getScopedClient(request),
     runner: manager.getRunner(),
     toolRegistry,
