+++ x-pack/platform/plugins/shared/agent_builder_platform/server/tools/index.ts
@@ -21,7 +21,6 @@ import { indexExplorerTool } from './index_explorer';
 import { generateEsqlTool } from './generate_esql';
 import { executeEsqlTool } from './execute_esql';
 import { searchTool } from './search';
-import { apiTools } from './api';
 
 export const registerTools = ({
   coreSetup,
@@ -42,7 +41,6 @@ export const registerTools = ({
     indexExplorerTool(),
     productDocumentationTool(coreSetup),
     integrationKnowledgeTool(coreSetup),
-    ...apiTools(coreSetup),
   ];
 
   tools.forEach((tool) => {
